+++ lib/audio_defect_detector.dart
@@ -21,3 +21,4 @@ export 'src/analyser.dart';
 export 'src/math_utils.dart';
 export 'src/models.dart';
 export 'src/pcm_decoder.dart';
+export 'src/version.dart';